+++ crates/ronami-core/src/types/chat_member.rs
@@ -776,6 +776,7 @@ mod tests {
                 language_code: Some("en".to_string()),
                 is_premium: false,
                 added_to_attachment_menu: false,
+                can_manage_bots: false,
             },
             kind: ChatMemberKind::Administrator(Administrator {
                 custom_title: None,
@@ -843,6 +844,7 @@ mod tests {
                 language_code: Some("en".to_string()),
                 is_premium: false,
                 added_to_attachment_menu: false,
+                can_manage_bots: false,
             },
             kind: ChatMemberKind::Restricted(Restricted {
                 tag: None,