+++ crates/ronami-core/src/types/me.rs
@@ -86,6 +86,7 @@ mod tests {
                 language_code: None,
                 is_premium: false,
                 added_to_attachment_menu: false,
+                can_manage_bots: false,
             },
             can_join_groups: false,
             can_read_all_group_messages: false,