+++ crates/ronami/src/dispatching/filter_ext.rs
@@ -149,6 +149,10 @@ define_message_ext! {
     (filter_web_app_data, Message::web_app_data),
     (filter_chat_owner_left, Message::chat_owner_left),
     (filter_chat_owner_changed, Message::chat_owner_changed),
+    (filter_managed_bot_created, Message::managed_bot_created),
+    (filter_poll_option_added, Message::poll_option_added),
+    (filter_poll_option_deleted, Message::poll_option_deleted),
+    (filter_reply_to_poll_option_id, Message::reply_to_poll_option_id),
 }
 
 macro_rules! define_update_ext {
@@ -192,4 +196,5 @@ define_update_ext! {
     (filter_chat_join_request, UpdateKind::ChatJoinRequest, ChatJoinRequest),
     (filter_chat_boost, UpdateKind::ChatBoost, ChatBoost),
     (filter_removed_chat_boost, UpdateKind::RemovedChatBoost, RemovedChatBoost),
+    (filter_managed_bot, UpdateKind::ManagedBot, ManagedBot),
 }