+++ crates/ronami/src/dispatching/handler_description.rs
@@ -85,6 +85,7 @@ impl EventKind for Kind {
             ChatJoinRequest,
             ChatBoost,
             RemovedChatBoost,
+            ManagedBot,
         ]
         .into_iter()
         .map(Kind)
@@ -166,6 +167,7 @@ mod tests {
             ChatJoinRequest,
             ChatBoost,
             RemovedChatBoost,
+            ManagedBot,
         ];
 
         for update in allowed_updates_reference {
@@ -193,7 +195,8 @@ mod tests {
                 | BusinessMessage
                 | BusinessConnection
                 | EditedBusinessMessage
-                | DeletedBusinessMessages => {
+                | DeletedBusinessMessages
+                | ManagedBot => {
                     assert!(full_set.contains(&Kind(update)))
                 }
             }